+++ packages/form/index.js
@@ -1,5 +1,8 @@
 import WeuiForm from './src/form'
 import WeuiFormGroup from './src/group'
+import WeuiFormTip from './src/tip'
+import WeuiFormOpr from './src/opr'
+import WeuiFormExtra from './src/extra'
 import WeuiInput from './src/input'
 import WeuiRadio from './src/radio'
 import WeuiRadioGroup from './src/radio-group'
@@ -12,6 +15,9 @@ import WeuiSwitch from './src/switch'
 WeuiForm.install = vue => {
   vue.component(WeuiForm.name, WeuiForm)
   vue.component(WeuiFormGroup.name, WeuiFormGroup)
+  vue.component(WeuiFormTip.name, WeuiFormTip)
+  vue.component(WeuiFormOpr.name, WeuiFormOpr)
+  vue.component(WeuiFormExtra.name, WeuiFormExtra)
   vue.component(WeuiInput.name, WeuiInput)
   vue.component(WeuiRadio.name, WeuiRadio)
   vue.component(WeuiRadioGroup.name, WeuiRadioGroup)
@@ -24,6 +30,9 @@ WeuiForm.install = vue => {
 export {
   WeuiForm,
   WeuiFormGroup,
+  WeuiFormTip,
+  WeuiFormOpr,
+  WeuiFormExtra,
   WeuiInput,
   WeuiRadio,
   WeuiRadioGroup,