+++ packages/form/src/extra.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-form__extra-area">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiFormExtra'
+}
+</script>