+++ packages/form/src/form.vue
@@ -4,10 +4,7 @@
       <h2 class="weui-form__title">{{ title }}</h2>
       <div class="weui-form__desc">{{ desc }}</div>
     </div>
-    <div class="weui-form__control-area">
-      <slot></slot>
-    </div>
-    <slot name="footer"></slot>
+    <slot></slot>
   </div>
 </template>
 <script>
@@ -19,10 +16,6 @@ export default {
       default: null,
     },
     desc: {
-      type: String,
-      default: null,
-    },
-    groupTitle: {
       type: String,
       default: null,
     },