+++ packages/form/src/opr.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-form__opr-area">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiFormOpr'
+}
+</script>