+++ examples/docs/changelog.md
@@ -5,3 +5,14 @@
 `2020-12-08`
 
 - 完成 20+ 对应 weui 的组件
+
+### progress
+
+#### Bug Fixes
+
+- button: 修复disabled变更时没有变更样式的问题
+
+#### Features
+
+- input: 支持v-model双向绑定
+- 增加 `form-tip` `form-opr` `from-extra` 组件